emulator base can be used with real navigation
