+++ cob_hardware_emulation/CMakeLists.txt
@@ -7,6 +7,6 @@ catkin_package()
 
 catkin_install_python(PROGRAMS
   scripts/emulation_follow_joint_trajectory.py
-  scripts/emulation_move_base.py
+  scripts/emulation_base.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
